+++ .gitmodules
@@ -1,0 +1,6 @@
+[submodule "evolve_pro"]
+	path = evolve_pro
+	url = https://github.com/mat10d/EvolvePro.git
+[submodule "external/evolvepro"]
+	path = external/evolvepro
+	url = https://github.com/mat10d/EvolvePro.git
