+++ README.md
@@ -1,1 +1,18 @@
 # ES activity prediction using Deep Learning
+
+## Installation
+
+```bash
+conda create -n bio python=3.10
+conda activate
+
+git clone https://github.com/sma1-rmarud/protein_demo.git
+cd protein_demo
+
+pip install -e .
+```
+
+## Usage
+```bash
+./run_gui.sh
+```
