+++ run_gui.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python -m biogui.app
